+++ Twiterizer.Framework/TwitterRequest.cs
@@ -232,13 +232,13 @@ namespace Twitterizer.Framework
 			if (Element == null) return null;
 
 			//Mon May 12 15:56:07 +0000 2008
-			Status.ID = int.Parse(Element["id"].InnerText);
+			Status.ID = long.Parse(Element["id"].InnerText);
 			Status.Created = ParseDateString(Element["created_at"].InnerText);
 			Status.Text = Element["text"].InnerText;
 			Status.Source = Element["source"].InnerText;
 			Status.IsTruncated = bool.Parse(Element["truncated"].InnerText);
 			if (Element["in_reply_to_status_id"].InnerText != string.Empty)
-				Status.InReplyToStatusID = int.Parse(Element["in_reply_to_status_id"].InnerText);
+				Status.InReplyToStatusID = long.Parse(Element["in_reply_to_status_id"].InnerText);
 			if (Element["in_reply_to_user_id"].InnerText != string.Empty)
 				Status.InReplyToUserID = int.Parse(Element["in_reply_to_user_id"].InnerText);
